backend/pyproject.toml
[project]
name = "jobalert-ai"
version = "1.0.0"
description = "AI-powered job matching platform with tailored resume generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-multipart==0.0.12",
    "motor==3.6.0",
    "beanie==1.26.0",
    "PyJWT==2.9.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "celery==5.4.0",
    "redis==5.2.0",
    "anthropic==0.39.0",
    "openai==1.58.1",
    "weasyprint==62.3",
    "sendgrid==6.11.0",
    "httpx==0.27.2",
    "beautifulsoup4==4.12.3",
    "feedparser==6.0.11",
    "lxml==5.3.0",
    "pydantic-settings==2.6.0",
    "python-dotenv==1.0.1",
    "boto3==1.35.36",
    "stripe==11.1.0",
    "sentry-sdk[fastapi]==2.17.0",
    "email-validator>=2.3.0",
    "reportlab>=4.4.10",
    "mongoengine>=0.29.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
